security: add support for master passphrase
- Add support for change master passphrase in NVM Protocol
- Add support for master erase in NVM Protocol
- Add support for change master passphrase in nvm api
- Add MasterPassphraseEnabled field to show dimm/HII detailed
- Add master passphrase to supported to show system capabilities

nvm_api 3.1.0:
 * add nvm_set_master_passphrase function
 * add master_passphrase_capable to security capabilities
 * add master_passphrase_enabled to device_discovery

Master passphrase is not supported in the CLI

+++ DcpmPkg/driver/Core/Diagnostics/SecurityDiagnostic.c
@@ -32,7 +32,7 @@ RunSecurityDiagnostics(
   EFI_STATUS ReturnCode = EFI_INVALID_PARAMETER;
   CHAR16 *pInconsistentSecurityStatesStr = NULL;
   UINT8 DimmSecurityState = 0;
-  UINT8 SecurityFlag = 0;
+  UINT32 SecurityFlag = 0;
   UINT8 SecurityStateCount[SECURITY_STATES_COUNT];
   BOOLEAN InconsistencyFlag = FALSE;
   UINT8 Index = 0;
